brokerage_service/pyproject.toml
[project]
name = "brokerage-service"
version = "0.1.0"
description = "A Model Context Protocol server providing IBKR brokerage access"
requires-python = ">=3.10"
dependencies = [
 "aiohttp>=3.11.8",
 "ib-insync>=0.9.86",
 "mcp[cli]>=1.0.0",
 "nest-asyncio>=1.6.0",
 "pydantic>=2.10.2",
 "python-dotenv>=1.0.1",
 "pytz>=2024.2",
 "requests>=2.32.3",
 "retry>=0.9.2",
 "setuptools>=75.6.0",
]

[[project.authors]]
name = "Itay Verkh"
email = "itay1542@gmail.com"

[project.scripts]
brokerage_service = "src.server:main"
